Update ton-blockchain.md
Under the workchain ID of the page - Smart Contract Addresses - https://docs.ton.org/learn/overviews/addresses#workchain-id-and-account-id

It explain as we have 32 bit workchain_id there is a possibility of 2^32 unique workchains, but this page says 2^30, which I felt is a typo(human error). 

If not a error can you please add some description to justify the delta in these numbers.