Update processing.md
`send_raw_msg` docs says that `mode = 1` is "pay fees separately from the message value". But in current document, it says that fees should be deducted from the message value, while also saying 'exactly 1 Toncoin is sent back'. 

I think that correct is 'mode=0' so that fees are deducted from that 1 TON, sorry if im wrong, i'm new to this stuff.