Update explorers-in-ton.mdx
The OKX TON Explorer is a powerful and user-friendly tool designed to provide comprehensive data insights into the TON (The Open Network) blockchain. It supports features such as address data, Jetton analysis, and NFT collection analysis, TON DNS, making it a valuable resource for both beginners and developers. By adding it to the list of third-party block explorers in the official documentation, users will have access to a broader range of tools to explore and interact with the TON ecosystem.

Add OKX TON Explorer introduction in Explorers in TON